Añado las funciones de wordcloud, folium y las corr de la target
